# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetConnect API
  description: |
    API REST para PetConnect — plataforma de mascotas que conecta dueños,
    veterinarios y comunidad. Incluye feed social, ranking semanal, InstaPet
    (red social por mascota), mascotas perdidas y adopciones.
  version: 0.1.0
  contact:
    name: PetConnect Team
servers:
  - url: http://localhost:8001/api/v1
    description: Servidor local de desarrollo (Hono)
  - url: https://api.petconnect.app/api/v1
    description: Producción
security:
  - bearerAuth: []
paths:
...
  /pets:
    get:
      tags: [Pets]
      summary: Listar mascotas
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: species
          in: query
          schema:
            type: string
          description: Filtrar por especie
        - name: owner_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrar por dueño
      responses:
        "200":
          description: Lista paginada de mascotas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPets"
    post:
      tags: [Pets]
      summary: Registrar nueva mascota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePetRequest"
      responses:
        "201":
          description: Mascota creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
        "422":
          $ref: "#/components/responses/ValidationError"
  /pets/{pet_id}:
    get:
      tags: [Pets]
      summary: Obtener mascota por ID
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "200":
          description: Detalle de mascota
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Pets]
      summary: Actualizar mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePetRequest"
      responses:
        "200":
          description: Mascota actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
    delete:
      tags: [Pets]
      summary: Eliminar mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "204":
          description: Mascota eliminada
        "404":
          $ref: "#/components/responses/NotFound"

  # --- Vet Visits ---
  /pets/{pet_id}/vet-visits:
    get:
      tags: [VetVisits]
      summary: Listar visitas al veterinario de una mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Lista paginada de visitas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedVetVisits"
    post:
      tags: [VetVisits]
      summary: Registrar visita al veterinario
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVetVisitRequest"
      responses:
        "201":
          description: Visita registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetVisit"
        "422":
          $ref: "#/components/responses/ValidationError"
  /pets/{pet_id}/vet-visits/{visit_id}:
    put:
      tags: [VetVisits]
      summary: Actualizar visita al veterinario
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/VisitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVetVisitRequest"
      responses:
        "200":
          description: Visita actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetVisit"
    delete:
      tags: [VetVisits]
      summary: Eliminar visita al veterinario
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/VisitId"
      responses:
        "204":
          description: Visita eliminada

  # --- Pet Events (Health Tracking) ---
  /pets/{pet_id}/events:
    get:
      tags: [PetEvents]
      summary: Listar eventos de health tracking de una mascota
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Lista paginada de eventos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPetEvents"
    post:
      tags: [PetEvents]
      summary: Registrar evento de health tracking
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePetEventRequest"
      responses:
        "201":
          description: Evento registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetEvent"
        "422":
          $ref: "#/components/responses/ValidationError"
  /pets/{pet_id}/events/{event_id}:
    put:
      tags: [PetEvents]
      summary: Actualizar evento
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePetEventRequest"
      responses:
        "200":
          description: Evento actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetEvent"
    delete:
      tags: [PetEvents]
      summary: Eliminar evento
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/EventId"
      responses:
        "204":
          description: Evento eliminado

  # =========================================================================
  # FEED (Social General)
  # =========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de Supabase Auth

  # ------------------------------------------------------------------
  # Parameters (reusables)
  # ------------------------------------------------------------------
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Número de página
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Resultados por página
    PetId:
      name: pet_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    VisitId:
      name: visit_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EventId:
      name: event_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ------------------------------------------------------------------
  # Schemas
  # ------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Recurso no encontrado"
      required: [detail]
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        pages:
          type: integer
          example: 3

    # --- Auth ---
...
    Pet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
          example: "Luna"
        species:
          type: string
          example: "Perro"
        breed:
          type: string
          nullable: true
          example: "Golden Retriever"
        age:
          type: integer
          nullable: true
          example: 24
        weight:
          type: number
          nullable: true
          example: 25.5
        photo_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePetRequest:
      type: object
      properties:
        name:
          type: string
          example: "Luna"
        species:
          type: string
          example: "Perro"
        breed:
          type: string
        age:
          type: integer
        weight:
          type: number
        photo_url:
          type: string
          format: uri
        bio:
          type: string
      required: [name, species]
    UpdatePetRequest:
      type: object
      properties:
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        age:
          type: integer
        weight:
          type: number
        photo_url:
          type: string
          format: uri
        bio:
          type: string
    PaginatedPets:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Pet"

    # --- Vet Visit ---
    VetVisit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pet_id:
          type: string
          format: uuid
        vet_name:
          type: string
          example: "Dr. Martínez"
        visit_date:
          type: string
          format: date
          example: "2026-05-15"
        reason:
          type: string
          example: "Revisión anual"
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateVetVisitRequest:
      type: object
      properties:
        vet_name:
          type: string
          example: "Dr. Martínez"
        visit_date:
          type: string
          format: date
        reason:
          type: string
        notes:
          type: string
      required: [vet_name, visit_date, reason]
    PaginatedVetVisits:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/VetVisit"

    # --- Pet Event (Health Tracking) ---
    PetEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pet_id:
          type: string
          format: uuid
        event_type:
          type: string
          enum: [vaccination, weight, deworming, medication, other]
          example: "vaccination"
        event_date:
          type: string
          format: date
          example: "2026-05-10"
        value:
          type: string
          nullable: true
          example: "Rabia"
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePetEventRequest:
      type: object
      properties:
        event_type:
          type: string
          enum: [vaccination, weight, deworming, medication, other]
        event_date:
          type: string
          format: date
        value:
          type: string
        notes:
          type: string
      required: [event_type, event_date]
    PaginatedPetEvents:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PetEvent"

    # --- Post (Feed) ---
...